# docs/swagger.yaml
basePath: /
definitions:
...
  request.ColorCustomizationRequest:
    properties:
      carId:
        type: string
      colorCategory:
        type: string
      colorCode:
        type: string
      colorName:
        type: string
      colorType:
        type: string
    type: object
...
  request.GetCarColorCategoriesRequest:
    properties:
      part:
        type: string
    type: object
  request.GetCarColorTypesRequest:
    properties:
      colorCategory:
        type: string
      part:
        type: string
    type: object
...
  request.InteriorCustomizeRequest:
    properties:
      carId:
        type: string
      colorCode:
        type: string
      colorName:
        type: string
    type: object
  request.LicenseRequest:
    properties:
      carId:
        type: string
      value:
        type: string
    type: object
...
  request.WheelCustomizeRequest:
    properties:
      carId:
        type: string
      colorCategory:
        type: string
      colorCode:
        type: string
      colorName:
        type: string
    type: object
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the doumentation for drag racing game
  title: Drag Racing
  version: "2.0"
paths:
...
  /car/customise/color:
    put:
      consumes:
      - application/json
      description: Customizes the color of a player's car.
      parameters:
      - description: The ID of the player
        in: header
        name: playerId
        required: true
        type: string
      - description: Color customization request object
        in: body
        name: colorReq
        required: true
        schema:
          $ref: '#/definitions/request.ColorCustomizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Color Customization
      tags:
      - Car-Customize
  /car/customise/interior:
    put:
      consumes:
      - application/json
      description: Customizes the interior of a player's car.
      parameters:
      - description: The ID of the player
        in: header
        name: playerId
        required: true
        type: string
      - description: Interior customization request object
        in: body
        name: interiorReq
        required: true
        schema:
          $ref: '#/definitions/request.InteriorCustomizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Interior Customization
      tags:
      - Car-Customize
  /car/customise/license:
    put:
      consumes:
      - application/json
      description: Customizes the license plate of a player's car.
      parameters:
      - description: The ID of the player
        in: header
        name: playerId
        required: true
        type: string
      - description: License plate customization request object
        in: body
        name: licenseRequest
        required: true
        schema:
          $ref: '#/definitions/request.LicenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: License Plate Customization
      tags:
      - Car-Customize
  /car/customise/wheels:
    put:
      consumes:
      - application/json
      description: Customizes the wheels of a player's car.
      parameters:
      - description: The ID of the player
        in: header
        name: playerId
        required: true
        type: string
      - description: Wheel customization request object
        in: body
        name: wheelReq
        required: true
        schema:
          $ref: '#/definitions/request.WheelCustomizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Wheel Customization
      tags:
      - Car-Customize
...
  /get-color-category:
    get:
      consumes:
      - application/json
      description: Retrieve a list of car color categories for a specific part
      parameters:
      - description: Car customization part
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/request.GetCarColorCategoriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get car color categories
      tags:
      - Car-Customize
  /get-color-type:
    get:
      consumes:
      - application/json
      description: Retrieve a list of car color types for a specific part and color
        category
      parameters:
      - description: Car color type like fluorescent,military
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/request.GetCarColorTypesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get car color types
      tags:
      - Car-Customize
  /get-colors:
    get:
      consumes:
      - application/json
      description: Retrieve a list of car color types for a specific part and color
        category
      parameters:
      - description: Car color type like fluorescent,military
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/request.GetCarColorTypesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get car color types
      tags:
      - Car-Customize
  /get-customization:
    get:
      consumes:
      - application/json
      description: Retrieve a list of car customization parts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get car customization parts
      tags:
      - Car-Customize
...
swagger: "2.0"